[Refactor]: 스플래쉬 스크린 분리
